backend/requirements.txt
fastapi==0.115.14
uvicorn[standard]
sqlalchemy~=2.0.41
mysql-connector-python~=9.3.0
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]~=1.7.4
bcrypt~=4.0.1
python-dotenv~=1.1.0
email_validator~=2.2.0
python-multipart~=0.0.20
google-adk~=1.5.0
python-dotenv~=1.1.0
python-multipart~=0.0.20
Authlib~=1.6.0
itsdangerous~=2.2.0
types-SQLAlchemy~=1.4.53.38
requests~=2.32.3
deprecated~=1.2.18
litellm~=1.72.1
asyncio~=3.4.3
fastmcp~=2.7.1
dataclasses~=0.6
uv~=0.7.12
chromadb==1.0.13
sentence-transformers>=2.2.2
apscheduler~=3.11.0
pymupdf>=1.23.0
matplotlib~=3.8.0
genanki~=0.13.0
pdf2image~=1.17.0
Pillow~=10.0.0